+++ zcash_primitives/src/legacy.rs
@@ -322,6 +322,11 @@ impl Script {
         Vector::write(&mut writer, &self.0, |w, e| w.write_u8(*e))
     }
 
+    /// Returns the length of this script as encoded (including the initial CompactSize).
+    pub fn serialized_size(&self) -> usize {
+        Vector::serialized_size(&self.0)
+    }
+
     /// Returns the address that this Script contains, if any.
     pub(crate) fn address(&self) -> Option<TransparentAddress> {
         if self.0.len() == 25