+++ zcash_primitives/CHANGELOG.md
@@ -11,6 +11,11 @@ and this library adheres to Rust's notion of
   - `impl From<TransparentKeyScope> for bip32::ChildNumber`
   - `impl From<NonHardenedChildIndex> for bip32::ChildNumber`
   - `impl TryFrom<bip32::ChildNumber> for NonHardenedChildIndex`
+- `zcash_primitives::legacy::Script::serialized_size`
+- `zcash_primitives::transaction::fees::transparent`:
+  - `InputSize`
+  - `InputView::serialized_size`
+  - `OutputView::serialized_size`
 
 ### Changed
 - MSRV is now 1.70.0.